Fixed mouse camera controls in shell.
Added UDModel::GetUDMatrix().
main.cpp now loads St Peters instead of DirCube.
Increased the window size in window.qml to 960 by 540.
Added TextFields to the toolbar to display the position and orientation of the camera. (Note these are temporary).
viewer.cpp sets the cameras position to be the translation component of the udmodels matrix.